change security config
