+++ apigateway/src/main/java/de/muenchen/isi/configuration/SecurityDisabledConfiguration.java
@@ -7,15 +7,20 @@ package de.muenchen.isi.configuration;
 import org.springframework.context.annotation.Bean;
 import org.springframework.context.annotation.Configuration;
 import org.springframework.context.annotation.Profile;
+import org.springframework.security.config.Customizer;
 import org.springframework.security.config.web.server.ServerHttpSecurity;
 import org.springframework.security.web.server.SecurityWebFilterChain;
 
 @Configuration
 @Profile("no-security")
-public class NoSecurityConfiguration {
+public class SecurityDisabledConfiguration {
 
     @Bean
-    public SecurityWebFilterChain springSecurityFilterChain(ServerHttpSecurity http) {
-        return http.authorizeExchange().anyExchange().permitAll().and().cors().and().csrf().disable().build();
+    public SecurityWebFilterChain springSecurityFilterChain(final ServerHttpSecurity http) {
+        return http
+            .authorizeExchange(exchange -> exchange.anyExchange().permitAll())
+            .cors(Customizer.withDefaults())
+            .csrf(ServerHttpSecurity.CsrfSpec::disable)
+            .build();
     }
 }
